move poller to hcore
